Module Opium

module Context : sig ... end

A context holds heterogeneous value and is passed to the requests or responses.

module Headers : sig ... end

Header Fields

module Cookie : sig ... end

Cookie management for both requests and responses.

module Method : sig ... end

Request Method

module Version : sig ... end

Protocol Version

module Status : sig ... end

Response Status Codes

module Body : sig ... end

Represents an HTTP request or response body.

module Request : sig ... end

Module to create and work with HTTP requests.

module Response : sig ... end

Module to create and work with HTTP responses.

module App : sig ... end

An opium app provides a set of convenience functions and types to construct a rock app.

module Route : sig ... end

Expression that represent a target or multiple

module Router : sig ... end

module Handler : sig ... end

Collection of handlers commonly used to build Opium applications

module Middleware : sig ... end

Collection of middlewares commonly used to build Opium applications.