Module Opium.Status

Response Status Codes

The status-code element is a three-digit integer code giving the result of the attempt to understand and satisfy the request.

See RFC72316 for more details.

type informational = [

| `Continue

| `Switching_protocols

]

The 1xx (Informational) class of status code indicates an interim response for communicating connection status or request progress prior to completing the requested action and sending a final response. See RFC72316.2 for more details.

type successful = [

| `OK

| `Created

| `Accepted

| `Non_authoritative_information

| `No_content

| `Reset_content

| `Partial_content

]

The 2xx (Successful) class of status code indicates that the client's request was successfully received, understood, and accepted. See RFC72316.3 for more details.

type redirection = [

| `Multiple_choices

| `Moved_permanently

| `Found

| `See_other

| `Not_modified

| `Use_proxy

| `Temporary_redirect

]

The 3xx (Redirection) class of status code indicates that further action needs to be taken by the user agent in order to fulfill the request. See RFC72316.4 for more details.

type client_error = [

| `Bad_request

| `Unauthorized

| `Payment_required

| `Forbidden

| `Not_found

| `Method_not_allowed

| `Not_acceptable

| `Proxy_authentication_required

| `Request_timeout

| `Conflict

| `Gone

| `Length_required

| `Precondition_failed

| `Payload_too_large

| `Uri_too_long

| `Unsupported_media_type

| `Range_not_satisfiable

| `Expectation_failed

| `Upgrade_required

| `I_m_a_teapot

| `Enhance_your_calm

]

The 4xx (Client Error) class of status code indicates that the client seems to have erred. See RFC72316.5 for more details.

type server_error = [

| `Internal_server_error

| `Not_implemented

| `Bad_gateway

| `Service_unavailable

| `Gateway_timeout

| `Http_version_not_supported

]

The 5xx (Server Error) class of status code indicates that the server is aware that it has erred or is incapable of performing the requested method. See RFC72316.6 for more details.

type standard = [

| informational

| successful

| redirection

| client_error

| server_error

]

The status codes defined in the HTTP 1.1 RFCs

type t = [

| standard

| `Code of int

]

The standard codes along with support for custom codes.

val default_reason_phrase : t -> string

default_reason_phrase standard is the example reason phrase provided by RFC7231 for the t status code. The RFC allows servers to use reason phrases besides these in responses.

val long_reason_phrase : t -> string

long_reason_phrase standard is an explanation of the the t status code.

val to_code : t -> int

to_code t is the integer representation of t.

val of_code : int -> t

of_code i is the t representation of i. of_code raises Failure if i is not a positive three-digit number.

val unsafe_of_code : int -> t

unsafe_of_code i is equivalent to of_code i, except it accepts any positive code, regardless of the number of digits it has. On negative codes, it will still raise Failure.

val is_informational : t -> bool

is_informational t is true iff t belongs to the Informational class of status codes.

val is_successful : t -> bool

is_successful t is true iff t belongs to the Successful class of status codes.

val is_redirection : t -> bool

is_redirection t is true iff t belongs to the Redirection class of status codes.

val is_client_error : t -> bool

is_client_error t is true iff t belongs to the Client Error class of status codes.

val is_server_error : t -> bool

is_server_error t is true iff t belongs to the Server Error class of status codes.

val is_error : t -> bool

is_error t is true iff t belongs to the Client Error or Server Error class of status codes.

Utilities

val to_string : t -> string

to_string t returns a string representation of the status t.

val of_string : string -> t

of_string s returns a status from its string representation s.

val sexp_of_t : t -> Sexplib0.Sexp.t

sexp_of_t t converts the request t to an s-expression

val pp : Stdlib.Format.formatter -> t -> unit

pp formats the request t as an s-expression

val pp_hum : Stdlib.Format.formatter -> t -> unit

pp_hum formats the request t as a standard HTTP request