Module Opium__.Route

type t

type matches = {

params : (string * string) list;

splat : string list;

}

val sexp_of_matches : matches -> Sexplib0.Sexp.t

sexp_of_t matches converts the matches matches to an s-expression

val of_string : string -> t

of_string s returns a route from its string representation s.

val to_string : t -> string

to_string t returns a string representation of the route t.

val match_url : t -> string -> matches option

match_url t url return the matches of the url url for the route t, or None if the url does not match.