Module Opium__.Version

type t = Httpaf.Version.t = {

major : int;

minor : int;

}

val compare : t -> t -> int

compare x y returns 0 if version x is equal to version y, a negative integer if version x is less than version y, and a positive integer if version x is greater than version y.

val to_string : t -> string

to_string t returns a string representation of the version t.

val of_string : string -> t

of_string s returns a version from its string representation s.

Utilities

val sexp_of_t : t -> Sexplib0.Sexp.t

sexp_of_t t converts the request t to an s-expression

val pp : Stdlib.Format.formatter -> t -> unit

pp formats the request t as an s-expression

val pp_hum : Stdlib.Format.formatter -> t -> unit

pp_hum formats the request t as a standard HTTP request